NET-85: moved bootstrap_controller to bootstrap package, fixed compilation
