+++ network/controller/bootstrap/bootstrap.go
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package controller
+package bootstrap
 
 import (
 	"context"
